Moving smashed sum interfaces to correct package #326
