+++ lisa/lisa-analyses/src/main/java/it/unive/lisa/analysis/string/CharInclusion.java
@@ -3,6 +3,7 @@ package it.unive.lisa.analysis.string;
 import it.unive.lisa.analysis.Lattice;
 import it.unive.lisa.analysis.SemanticException;
 import it.unive.lisa.analysis.SemanticOracle;
+import it.unive.lisa.analysis.combination.SmashedSumStringDomain;
 import it.unive.lisa.analysis.lattices.Satisfiability;
 import it.unive.lisa.program.cfg.ProgramPoint;
 import it.unive.lisa.symbolic.value.Constant;
