+++ lisa/lisa-analyses/src/main/java/it/unive/lisa/analysis/string/bricks/Bricks.java
@@ -3,8 +3,8 @@ package it.unive.lisa.analysis.string.bricks;
 import it.unive.lisa.analysis.Lattice;
 import it.unive.lisa.analysis.SemanticException;
 import it.unive.lisa.analysis.SemanticOracle;
+import it.unive.lisa.analysis.combination.SmashedSumStringDomain;
 import it.unive.lisa.analysis.lattices.Satisfiability;
-import it.unive.lisa.analysis.string.SmashedSumStringDomain;
 import it.unive.lisa.program.cfg.ProgramPoint;
 import it.unive.lisa.symbolic.value.Constant;
 import it.unive.lisa.symbolic.value.operator.binary.BinaryOperator;
