+++ lisa/lisa-analyses/src/main/java/it/unive/lisa/analysis/string/fsa/FSA.java
@@ -3,9 +3,9 @@ package it.unive.lisa.analysis.string.fsa;
 import it.unive.lisa.analysis.Lattice;
 import it.unive.lisa.analysis.SemanticException;
 import it.unive.lisa.analysis.SemanticOracle;
+import it.unive.lisa.analysis.combination.SmashedSumStringDomain;
 import it.unive.lisa.analysis.lattices.Satisfiability;
 import it.unive.lisa.analysis.numeric.Interval;
-import it.unive.lisa.analysis.string.SmashedSumStringDomain;
 import it.unive.lisa.program.cfg.ProgramPoint;
 import it.unive.lisa.symbolic.value.Constant;
 import it.unive.lisa.symbolic.value.operator.binary.BinaryOperator;
